+++ jgclark.Dashboard/src/demoData.js
@@ -1,7 +1,7 @@
 // @flow
 //-----------------------------------------------------------------------------
 // Demo data for Dashboard plugin (for v2.0.0+)
-// Last updated 2024-09-13 for v2.1.0.a11 by @jgclark
+// Last updated 2024-09-20 for v2.1.0.a12 by @jgclark
 //-----------------------------------------------------------------------------
 
 import moment from 'moment/min/moment-with-locales'
@@ -34,6 +34,9 @@ export const openTodayItems: Array<TSectionItem> = [
       priority: 0,
       content: 'before bed routine 🛌 22:00-22:30',
       rawContent: 'before bed routine 🛌 22:00-22:30',
+      indentLevel: 0,
+      isAChild: false,
+      hasChild: false,
     },
   },
   // $FlowIgnore[prop-missing] ID gets added later
@@ -46,6 +49,9 @@ export const openTodayItems: Array<TSectionItem> = [
       priority: 1,
       content: "reconcile bank statement @repeat(1m)",
       rawContent: "reconcile bank statement @repeat(1m)",
+      indentLevel: 0,
+      isAChild: false,
+      hasChild: false,
     }
   },
   // $FlowIgnore[prop-missing] ID gets added later
@@ -58,6 +64,9 @@ export const openTodayItems: Array<TSectionItem> = [
       priority: 0,
       content: 'check ==highlights==, `formatted` and ~~strike~~ text work OK',
       rawContent: 'check ==highlights==, `formatted` and ~~strike~~ text work OK',
+      indentLevel: 0,
+      isAChild: false,
+      hasChild: false,
     },
   },
   // $FlowIgnore[prop-missing] ID gets added later
@@ -70,6 +79,9 @@ export const openTodayItems: Array<TSectionItem> = [
       priority: 0,
       content: "morning @home 🏠 routine 6:30AM",
       rawContent: "morning @home 🏠 routine 6:30AM",
+      indentLevel: 0,
+      isAChild: false,
+      hasChild: false,
     }
   },
   // $FlowIgnore[prop-missing] ID gets added later
@@ -82,6 +94,9 @@ export const openTodayItems: Array<TSectionItem> = [
       priority: 0,
       content: "morning #work checklist 7:30AM",
       rawContent: "morning #work checklist 7:30AM",
+      indentLevel: 0,
+      isAChild: false,
+      hasChild: false,
     }
   },
 ]
@@ -97,6 +112,9 @@ export const refTodayItems: Array<TSectionItem> = [
       priority: 0,
       content: 'Edit video from CFL visit https://bcfd.org.uk 14:30-15:30',
       rawContent: 'Edit video from CFL visit https://bcfd.org.uk 14:30-15:30',
+      indentLevel: 0,
+      isAChild: false,
+      hasChild: false,
       blockId: '^wazhht',
     },
   },
@@ -111,6 +129,8 @@ export const refTodayItems: Array<TSectionItem> = [
       priority: 0,
       content: 'Update display board 08:00-09:00',
       rawContent: 'Update display board 08:00-09:00',
+      indentLevel: 0,
+      isAChild: false,
       hasChild: true,
     },
   },
@@ -125,6 +145,9 @@ export const refTodayItems: Array<TSectionItem> = [
       priority: 0,
       content: "Pay in cash from cafe 2:30PM",
       rawContent: "Pay in cash from cafe 2:30PM",
+      indentLevel: 0,
+      isAChild: false,
+      hasChild: false,
     }
   },
 ]
@@ -145,6 +168,8 @@ export const openYesterdayParas: Array<TSectionItem> = [
       priority: 4,
       content: '>> #editvideo from CFL visit @work',
       rawContent: '>> #editvideo from CFL visit @work',
+      indentLevel: 0,
+      isAChild: false,
       hasChild: true,
     },
   },
@@ -158,6 +183,8 @@ export const openYesterdayParas: Array<TSectionItem> = [
       priority: 0,
       content: 'update contract for [[Staff Induction (SW)]] following review comments',
       rawContent: 'update contract for [[Staff Induction (SW)]] following review comments',
+      indentLevel: 0,
+      isAChild: false,
       hasChild: true,
     },
   },
@@ -174,6 +201,9 @@ export const refYesterdayParas: Array<TSectionItem> = [
       priority: 1,
       content: 'write 5/3 sermon >2023-03-02',
       rawContent: '* write 5/3 sermon >2023-03-02',
+      indentLevel: 0,
+      isAChild: false,
+      hasChild: false,
       changedDate: new Date('2023-03-02T00:00:00.000Z'),
     },
   },
@@ -187,10 +217,11 @@ export const refYesterdayParas: Array<TSectionItem> = [
       title: 'Services',
       content: '  create 5/3 sermon presentation >2023-03-02',
       rawContent: '  * create 5/3 sermon presentation >2023-03-02',
+      indentLevel: 1,
+      isAChild: true,
+      hasChild: false,
       changedDate: new Date('2023-03-02T00:00:00.000Z'),
       priority: 1,
-      hasChild: false,
-      isAChild: true,
     },
   },
   // $FlowIgnore[prop-missing] ID gets added later
@@ -204,8 +235,10 @@ export const refYesterdayParas: Array<TSectionItem> = [
       priority: 1,
       content: '! Respond on Repair Cafe things from last 2 meetings >today #win ^wazhht',
       rawContent: '* ! Respond on Repair Cafe things from last 2 meetings >today #win ^wazhht',
-      blockId: '^wazhht',
+      indentLevel: 0,
+      isAChild: false,
       hasChild: true,
+      blockId: '^wazhht',
     },
   },
 ]
@@ -224,6 +257,9 @@ export const openTomorrowParas: Array<TSectionItem> = [
       filename: thisFilename,
       content: 'Clear more of prayer room @staff >today ^q9jzj4',
       rawContent: 'Clear more of prayer room @staff >today ^q9jzj4',
+      indentLevel: 0,
+      isAChild: false,
+      hasChild: false,
       priority: 0,
       blockId: '^q9jzj4',
       changedDate: new Date('2023-03-02T00:00:00.000Z'),
@@ -246,6 +282,9 @@ export const openWeekParas: Array<TSectionItem> = [
       priority: 0,
       content: ' Get login for https://www.waverleyabbeyresources.org/resources-home/',
       rawContent: ' Get login for https://www.waverleyabbeyresources.org/resources-home/',
+      indentLevel: 0,
+      isAChild: false,
+      hasChild: false,
       changedDate: new Date('2023-02-27T00:00:00.000Z'),
     },
   },
@@ -258,10 +297,12 @@ export const openWeekParas: Array<TSectionItem> = [
       filename: thisFilename,
       content: '@church Get iPad working on ChurchOffice-staff for Drive, Printing @church ^bzlp1z',
       rawContent: '@church Get iPad working on ChurchOffice-staff for Drive, Printing @church ^bzlp1z',
+      indentLevel: 0,
+      isAChild: false,
+      hasChild: true,
       blockId: '^bzlp1z',
       priority: 0,
       changedDate: new Date('2023-02-27T00:00:00.000Z'),
-      hasChild: true,
     },
   },
 ]
@@ -276,6 +317,9 @@ export const refWeekParas: Array<TSectionItem> = [
       priority: 2,
       content: '!! Arrange EV charger repair',
       rawContent: '!! Arrange EV charger repair',
+      indentLevel: 0,
+      isAChild: false,
+      hasChild: false,
       changedDate: new Date('2023-02-27T00:00:00.000Z'),
     },
   },
@@ -289,6 +333,9 @@ export const refWeekParas: Array<TSectionItem> = [
       priority: 0,
       content: 'Test multi-part hashtags: #project/companyA and #one/two/three >2023-W09',
       rawContent: 'Test multi-part hashtags: #project/companyA and #one/two/three >2023-W09',
+      indentLevel: 0,
+      isAChild: false,
+      hasChild: false,
       changedDate: new Date('2023-02-27T00:00:00.000Z'),
     },
   },
@@ -303,6 +350,9 @@ export const refWeekParas: Array<TSectionItem> = [
       priority: 0,
       content: 'Re-plant two shrubs in new blue pots @home >2023-W09',
       rawContent: 'Re-plant two shrubs in new blue pots @home >2023-W09',
+      indentLevel: 0,
+      isAChild: false,
+      hasChild: false,
       changedDate: new Date('2023-02-27T00:00:00.000Z'),
     },
   },
@@ -317,6 +367,9 @@ export const refWeekParas: Array<TSectionItem> = [
       priority: 0,
       content: 'Backup Mac - with an arrow date >2023-W09< reference',
       rawContent: 'Backup Mac - with an arrow date >2023-W09< reference',
+      indentLevel: 0,
+      isAChild: false,
+      hasChild: false,
       changedDate: new Date('2023-02-27T00:00:00.000Z'),
     },
   },
@@ -336,6 +389,9 @@ export const openMonthParas: Array<TSectionItem> = [
       priority: 0,
       content: 'Investigate alternative milkman',
       rawContent: 'Investigate alternative milkman',
+      indentLevel: 0,
+      isAChild: false,
+      hasChild: false,
     },
   },
 ]
@@ -351,6 +407,9 @@ export const refMonthParas: Array<TSectionItem> = [
       priority: 1,
       content: 'Pay tax bill',
       rawContent: 'Pay tax bill',
+      indentLevel: 0,
+      isAChild: false,
+      hasChild: false,
     },
   },
 ]
@@ -379,6 +438,7 @@ export const demoTaggedParas: Array<TParagraph> = [
     filename: 'TEST/DEMOs/Test Project A.md',
     content: 'Open Deliveroo account #next #home',
     rawContent: 'Open Deliveroo account #next  #home',
+    indentLevel: 0,
     priority: 0,
   },
   // $FlowIgnore[prop-missing] ID gets added later
@@ -388,6 +448,7 @@ export const demoTaggedParas: Array<TParagraph> = [
     filename: 'TEST/DEMOs/Test Project A.md',
     content: '#home Checklist ⛔️ that should be filtered out #waiting #next',
     rawContent: '#home Checklist ⛔️ that should be filtered out #waiting #next',
+    indentLevel: 0,
     priority: 0,
   },
   // $FlowIgnore[prop-missing] ID gets added later
@@ -397,6 +458,7 @@ export const demoTaggedParas: Array<TParagraph> = [
     filename: 'TEST/DEMOs/Test Project A.md',
     content: '#home Future task ⛔️ that should be filtered out #next >2099-W09',
     rawContent: '#home Future task ⛔️ that should be filtered out #next >2099-W09',
+    indentLevel: 0,
     priority: 0,
   },
   // $FlowIgnore[prop-missing] ID gets added later
@@ -406,6 +468,7 @@ export const demoTaggedParas: Array<TParagraph> = [
     filename: thisFilename,
     content: '@work Get iPad working on ChurchOffice-staff for printing etc. @church ^bzlp1z',
     rawContent: '@work Get iPad working on ChurchOffice-staff for printing etc. @church ^bzlp1z',
+    indentLevel: 0,
     priority: 0,
     blockId: '^bzlp1z',
     hasChild: true,
@@ -417,12 +480,14 @@ export const demoTaggedParas: Array<TParagraph> = [
     filename: 'CCC Areas/Finance.md',
     content: 'Make expenses claim @work',
     rawContent: 'Make expenses claim @work',
+    indentLevel: 0,
     priority: 0,
   },
 ]
 
 //-----------------------------------------------------------
 // Project Notes to review
+// Note: These are all overloaded for ease
 export const nextProjectNoteItems: Array<TNote> = [
   // $FlowIgnore[prop-missing]
   {
@@ -430,7 +495,8 @@ export const nextProjectNoteItems: Array<TNote> = [
     title: 'Hearing Support at CCC',
     reviewInterval: "1m",
     percentComplete: 23,
-    lastProgressComment: "Checked our equipment and its OK; looking for acoustician"
+    lastProgressComment: "Checked our equipment and its OK; looking for acoustician",
+    nextReviewDays: 10,
   },
   // $FlowIgnore[prop-missing]
   {
@@ -438,12 +504,14 @@ export const nextProjectNoteItems: Array<TNote> = [
     title: 'Streaming Platform',
     reviewInterval: "1w",
     percentComplete: 82,
+    nextReviewDays: 6,
   },
   // $FlowIgnore[prop-missing]
   {
     filename: 'CCC Projects/Pastoral Cards.md',
     title: 'Pastoral Cards',
     reviewInterval: "2m",
+    nextReviewDays: 20,
   },
 ]
 
@@ -466,6 +534,7 @@ export function makeDummyOverdueItems(extension: string): Array<TParagraph> {
       filename: filename,
       content: content,
       rawContent: `${thisType === 'open' ? '*' : '+'} ${priorityPrefix}${content}`,
+      indentLevel: 0,
       type: thisType,
       note: {
         filename: filename,
@@ -498,6 +567,7 @@ export function makeDummyPriorityItems(extension: string): Array<TParagraph> {
       filename: filename,
       content: content,
       rawContent: `${thisType === 'open' ? '*' : '+'} ${priorityPrefix}${content}`,
+      indentLevel: 0,
       type: thisType,
       note: {
         filename: filename,