+++ jgclark.Dashboard/src/react/components/CalendarPicker.jsx
@@ -37,23 +37,26 @@ const CalendarPicker = ({ onSelectDate, numberOfMonths = 2, startingSelectedDate
   //     '--rdp-cell-size': '20px', // Size down the calendar cells (default is 40px)
 
   // TODO: looks like these could all move to CalendarPicker.css?
-  const calendarStyles = { /* note: the non-color styles are set in CalendarPicker.css */
-    container: { border: '1px solid #ccc', marginTop: '0px', paddingTop: '0px' },
-    caption: { color: 'var(--tint-color)' },
-    navButtonPrev: { color: 'var(--tint-color)' },
-    navButtonNext: { color: 'var(--tint-color)' },
-    weekdays: { backgroundColor: 'var(--bg-main-color)' },
-    weekday: { fontWeight: 'bold' },
-    weekend: { backgroundColor: 'var(--bg-alt-color)' },
-    week: { color: '#333' },
-    day: { color: 'var(--fg-main-color)' },
-    today: { color: 'var(--hashtag-color)', backgroundColor: 'var(--bg-alt-color)' },
-    selected: { color: 'var(--tint-color)', backgroundColor: 'var(--bg-alt-color)' },
-  }
+  /* note: the non-color styles are set in CalendarPicker.css */
+  // const calendarStyles = {
+  //   container: { border: '1px solid #ccc', marginTop: '0px', paddingTop: '0px' },
+  //   caption: { color: 'var(--tint-color)' },
+  //   navButtonPrev: { color: 'var(--tint-color)' },
+  //   navButtonNext: { color: 'var(--tint-color)' },
+  //   weekdays: { backgroundColor: 'var(--bg-main-color)' },
+  //   weekday: { fontWeight: 'bold' },
+  //   weekend: { backgroundColor: 'var(--bg-alt-color)' },
+  //   week: { color: '#333' },
+  //   day: { color: 'var(--fg-main-color)' },
+  //   today: { color: 'var(--hashtag-color)', backgroundColor: 'var(--bg-alt-color)' },
+  //   selected: { color: 'var(--tint-color)', backgroundColor: 'var(--bg-alt-color)' },
+  // }
 
   return (
     <>
-      <button className="PCButton" onClick={toggleDatePicker}>
+      <button className="PCButton"
+        title="Open calendar to pick a specific day"
+        onClick={toggleDatePicker}>
         <i className="fa-solid fa-calendar-alt pad-left pad-right"></i>
       </button>
       {isOpen && (
@@ -65,7 +68,7 @@ const CalendarPicker = ({ onSelectDate, numberOfMonths = 2, startingSelectedDate
             numberOfMonths={numberOfMonths}
             required
             fixedHeight
-            styles={calendarStyles}
+            // styles={calendarStyles}
             className="calendarPickerCustom"
           />
         </div>