+++ jgclark.Dashboard/src/react/components/NoProjects.jsx
@@ -1,10 +1,10 @@
 // @flow
 //--------------------------------------------------------------------------
 // Dashboard React component to show there are no Projects overdue for review. Called by ItemRow.
-// Last updated 2024-08-13 for v2.2.0.a7 by @jgclark
+// Last updated 2024-08-13 for v2.2.0.a7+ by @jgclark
 //--------------------------------------------------------------------------
 
-import { type Node } from 'react'
+import React, { type Node } from 'react'
 
 /**
  * Component for displaying a message when there are no tasks.