+++ jgclark.Dashboard/src/react/components/NoTasks.jsx
@@ -1,10 +1,10 @@
 // @flow
 //--------------------------------------------------------------------------
 // Dashboard React component to show there are no tasks for today. Called by ItemRow or ItemContent.
-// Last updated 24.6.2024 for v2.0.0-b14 by @jgclark
+// Last updated 24.6.2024 for v2.0.0-b14+ by @jgclark
 //--------------------------------------------------------------------------
 
-import {type Node} from 'react'
+import React, { type Node } from 'react'
 
 /**
  * Component for displaying a message when there are no tasks.