+++ jgclark.Dashboard/src/react/components/TextComponent.jsx
@@ -2,7 +2,7 @@
 //----------------------------------------------------------
 // Dashboard React component to show a piece of text in the Settings and Dropdowns
 // TODO: This needs more thought on classNames of what it returns, as it is now being used by Settings more than Dropdowns (if at all).
-// Last updated 2024-08-15 for v2.1.0.a8 by @jgclark
+// Last updated 2024-09-21 for v2.1.0.a12 by @jgclark
 //----------------------------------------------------------
 
 import React from 'react'
@@ -10,20 +10,20 @@ import React from 'react'
 type TextComponentProps = {
   label: string,
   textType: 'title' | 'description' | 'separator' | 'header',
-  description?: string,
 };
 
-const TextComponent = ({ label, textType, description }: TextComponentProps): React$Node => {
+const TextComponent = ({ label, textType }: TextComponentProps): React$Node => {
   switch (textType) {
     case 'title':
       return <div className="dropdown-title">{label}</div>
     case 'header':
+      // TODO: despite this, it still shows 'description' field twice
       return (
         <>
           <div className="ui-heading">{label}</div>
-          {description && (
+          {/* {description && (
             <p className="item-description">{description}</p>
-          )}
+          )} */}
         </>
       )
     case 'description':